+++ apps/db/.loader.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-   'sequelize.js': {
-      level: 11
-   }
-}